% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/raqui.R
\name{ra_get_page__}
\alias{ra_get_page__}
\title{low level get page}
\usage{
ra_get_page__(cons, arq, message = "Arquivo obtido com sucesso!")
}
\arguments{
\item{cons}{url da p치gina consultada (excluido o dominio)}

\item{arq}{arquivo em que deseja salvar a p치gina}

\item{message}{mensagem que deseja exibir se o arquivo for salvo com sucesso}
}
\description{
Obtem o c칩digo html renderizado da p치gina do reclame aqui
pesquisada e salva em um arquivo
}

