% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/raqui.R
\name{ra_next_page}
\alias{ra_next_page}
\title{next page
busca o link da pr칩xima p치gina de reclamacoes}
\usage{
ra_next_page(page)
}
\description{
next page
busca o link da pr칩xima p치gina de reclamacoes
}

