% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/raqui.R
\name{ra_parse_page}
\alias{ra_parse_page}
\title{parse pagina
Faz o parse da página}
\usage{
ra_parse_page(page)
}
\arguments{
\item{page}{página em html do R (formato retornato por html(url))}
}
\description{
parse pagina
Faz o parse da página
}

