% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/raqui.R
\name{ra_print_captcha}
\alias{ra_print_captcha}
\title{função para imprimir o captcha recebido
não funciona!!!}
\usage{
ra_print_captcha(page, cookies)
}
\description{
função para imprimir o captcha recebido
não funciona!!!
}

